+++ server/src/api/forum/threads/h_list.go
@@ -1,4 +1,4 @@
-package forum
+package threads
 
 import (
 	"encoding/json"
